@@ -15,44 +15,50 @@
 	},
 	"private": true,
 	"dependencies": {
-		"@angular/animations": "^5.2.0",
-		"@angular/common": "^5.2.0",
-		"@angular/compiler": "^5.2.0",
-		"@angular/core": "^5.2.0",
-		"@angular/forms": "^5.2.0",
-		"@angular/http": "^5.2.0",
-		"@angular/platform-browser": "^5.2.0",
-		"@angular/platform-browser-dynamic": "^5.2.0",
-		"@angular/router": "^5.2.0",
-		"@fortawesome/fontawesome": "^1.1.4",
-		"angular-font-awesome": "^3.1.2",
-		"bootstrap": "^4.0.0",
-		"core-js": "^2.4.1",
-		"font-awesome": "^4.7.0",
+		"@angular/animations": "^6.1.7",
+		"@angular/cdk": "^6.4.7",
+		"@angular/common": "^6.1.7",
+		"@angular/compiler": "^6.1.7",
+		"@angular/core": "^6.1.7",
+		"@angular/flex-layout": "^6.0.0-beta.18",
+		"@angular/forms": "6.1.7",
+		"@angular/http": "^6.1.7",
+		"@angular/material": "^6.4.7",
+		"@angular/platform-browser": "^6.1.7",
+		"@angular/platform-browser-dynamic": "^6.1.7",
+		"@angular/router": "^6.1.7",
+		"bootstrap": "^4.1.3",
+		"core-js": "^2.5.7",
+		"imagesloaded": "^4.1.4",
 		"jquery": "^3.3.1",
-		"less": "^2.7.3",
-		"rxjs": "^5.5.6",
-		"zone.js": "^0.8.19"
+		"less": "^3.8.1",
+		"masonry-layout": "^4.2.2",
+		"ng-masonry-grid": "^1.3.0",
+		"npm": "^6.4.1",
+		"rxjs": "^6.3.2",
+		"rxjs-compat": "^6.3.2",
+		"zone.js": "^0.8.26"
 	},
 	"devDependencies": {
-		"@angular/cli": "1.6.5",
-		"@angular/compiler-cli": "^5.2.0",
-		"@angular/language-service": "^5.2.0",
-		"@types/jasmine": "~2.8.3",
-		"@types/jasminewd2": "~2.0.2",
-		"@types/node": "~6.0.60",
-		"codelyzer": "^4.0.1",
-		"jasmine-core": "~2.8.0",
+		"@angular-devkit/build-angular": "~0.8.0",
+		"@angular/cli": "^6.2.1",
+		"@angular/compiler-cli": "^6.1.7",
+		"@angular/language-service": "^6.1.7",
+		"@types/jasmine": "~2.8.8",
+		"@types/jasminewd2": "~2.0.3",
+		"@types/node": "~10.10.0",
+		"codelyzer": "^4.4.4",
+		"jasmine-core": "^3.2.1",
 		"jasmine-spec-reporter": "~4.2.1",
-		"karma": "~2.0.0",
+		"karma": "^3.0.0",
 		"karma-chrome-launcher": "~2.2.0",
 		"karma-cli": "~1.0.1",
-		"karma-coverage-istanbul-reporter": "^1.2.1",
-		"karma-jasmine": "~1.1.0",
-		"karma-jasmine-html-reporter": "^0.2.2",
-		"protractor": "~5.1.2",
-		"ts-node": "~4.1.0",
-		"tslint": "~5.9.1",
-		"typescript": "~2.5.3"
+		"karma-coverage-istanbul-reporter": "^2.0.4",
+		"karma-jasmine": "^1.1.2",
+		"karma-jasmine-html-reporter": "^1.3.1",
+		"protractor": "^5.4.1",
+		"ts-node": "^7.0.1",
+		"tslint": "^5.11.0",
+		"typescript": "^2.9.2"
 	}
 }
