@@ -4,24 +4,22 @@
 module.exports = function(config) {
 	config.set({
 		basePath: "",
-		frameworks: ["jasmine", "@angular/cli"],
+		frameworks: ["jasmine", "@angular-devkit/build-angular"],
 		plugins: [
 			require("karma-jasmine"),
 			require("karma-chrome-launcher"),
 			require("karma-jasmine-html-reporter"),
 			require("karma-coverage-istanbul-reporter"),
-			require("@angular/cli/plugins/karma")
+			require("@angular-devkit/build-angular/plugins/karma")
 		],
 		client: {
 			clearContext: false // leave Jasmine Spec Runner output visible in browser
 		},
 		coverageIstanbulReporter: {
-			reports: ["html", "lcovonly"],
+			dir: require('path').join(__dirname, 'coverage'), reports: ["html", "lcovonly"],
 			fixWebpackSourcePaths: true
 		},
-		angularCli: {
-			environment: "dev"
-		},
+		
 		reporters: ["progress", "kjhtml"],
 		port: 9876,
 		colors: true,
